refactor: Refactor code structure
Moved files to the correct directories
