+++ package.json
@@ -10,12 +10,11 @@
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
     "startnew": "nodemon",
-    "start:local": "docker-compose up && npx prisma migrate dev && nest start --watch",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
-    "start:dbformat": "npx prisma format",
-    "prisma:dev": "npx prisma migrate dev && npx prisma db seed",
+    "prisma:format": "npx prisma format",
+    "prisma:migrate": "npx prisma migrate dev && npx prisma db seed",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -26,7 +25,7 @@
 
   },
   "prisma": {
-    "seed": "ts-node src/scripts/seed"
+    "seed": "ts-node prisma/seed"
   },
   "dependencies": {
     "@darraghor/eslint-plugin-nestjs-typed": "^3.22.6",
@@ -39,7 +38,6 @@
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/swagger": "^6.3.0",
     "@prisma/client": "^4.15.0",
-    "@vercel/postgres": "^0.3.1",
     "bcrypt": "^5.1.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
@@ -84,7 +82,8 @@
   },
   "lint-staged": {
     "src/**/*.{js,ts}": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "src/**/*.ts": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix"
+    "src/**/*.ts": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "prisma:format": "npx prisma format"
   },
   "jest": {
     "moduleFileExtensions": [
