--- src/common/decorators/index.ts
@@ -1,3 +1,0 @@
-export * from '@app/src/common/decorators/get.current.user.decorator';
-export * from '@app/src/common/decorators/get.current.userId.decorator';
-export * from '@app/src/common/decorators/public.decorator';
