--- src/common/guards/index.ts
@@ -1,2 +1,0 @@
-export * from '@app/src/common/guards/at.guard';
-export * from '@app/src/common/guards/rt.guard';
