--- src/filters/index.ts
@@ -1,1 +1,0 @@
-export * from '@app/src/filters/prisma-client-exception.filter';
