+++ src/modules/auth/decorators/index.ts
@@ -1,0 +1,3 @@
+export * from '@app/src/modules/auth/decorators/get.current.user.decorator';
+export * from '@app/src/modules/auth/decorators/get.current.userId.decorator';
+export * from '@app/src/modules/auth/decorators/public.decorator';
