+++ src/modules/auth/guards/index.ts
@@ -1,0 +1,2 @@
+export * from '@app/src/modules/auth/guards/at.guard';
+export * from '@app/src/modules/auth/guards/rt.guard';
