+++ src/modules/prisma/filters/index.ts
@@ -1,0 +1,1 @@
+export * from '@app/src/modules/prisma/filters/prisma-client-exception.filter';
